server/requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
python-dotenv
requests
flask-cors
playwright